requirements.txt
Flask==2.0.3
black==19.10b0
flake8==4.0.1
isort==5.9.3
pytest==7.0.1
psycopg2==2.8.6
alembic==1.7.6
kafka-python==2.0.2
ipykernel==6.15.0
PyYAML==6.0
pandas==1.4.3
pydantic==1.9.1
timeout-decorator==0.5.0